requirements.txt
absl-py==2.1.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bokeh==3.3.3
branca==0.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deepmultilingualpunctuation==1.0.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.15.1
Flask==3.0.3
flatbuffers==24.3.25
folium==0.15.1
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.6.0
future==1.0.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.23.4
idna==3.6
intel-openmp==2021.4.0
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.4
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
mysql-connector-python==8.4.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
openai-whisper==20231117
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.3
referencing==0.33.0
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.4
stack-data==0.6.3
sympy==1.12.1
tbb==2021.12.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.0
tf_keras==2.16.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.41.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.1
validators==0.28.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xyzservices==2023.10.1
